refactor: bump to strict-types 1.4.x
