+++ src/rgb/structs.rs
@@ -2,8 +2,7 @@ use std::{collections::HashMap, str::FromStr};
 
 use bitcoin::Address;
 use bitcoin_scripts::address::AddressCompat;
-use bp::Outpoint;
-use rgb::{interface::OutpointFilter, RgbWallet, TerminalPath};
+use rgb::{RgbWallet, TerminalPath};
 
 use serde::{Deserialize, Serialize};
 
@@ -39,11 +38,4 @@ pub struct RgbAccount {
 pub struct AddressTerminal {
     pub address: AddressCompat,
     pub terminal: TerminalPath,
-}
-
-pub struct EmptyFilter {}
-impl OutpointFilter for EmptyFilter {
-    fn include_outpoint(&self, _outpoint: Outpoint) -> bool {
-        true
-    }
 }
